+++ src/plugin.ts
@@ -8,6 +8,7 @@ import { AutoClassPluginSettingsTab } from './settings';
 export class AutoClassPlugin extends Plugin {
   appliedClasses: ViewAppliedClasses[] = [];
   settings: AutoClassPluginSettings = DEFAULT_SETTINGS;
+
   async onload() {
     const savedData = await this.loadData();
     this.settings = Object.assign({}, DEFAULT_SETTINGS, savedData);
@@ -21,59 +22,92 @@ export class AutoClassPlugin extends Plugin {
     this.removeAllClasses();
   }
 
+  /**
+   * Layout change event handler. Removes old classes
+   * and applies new classes to the active view and
+   * any linked panes.
+   */
   handleLayoutChange(): void {
     const activeView = this.app.workspace.getActiveViewOfType(MarkdownView);
     if (activeView) {
-      console.log('REMOVING PREVIOUS');
-      this.removePreviousClasses(activeView);
-      let matches: ClassPath[] = [];
-      let container: Element;
-      if (this.isPreivewMode(activeView)) {
-        console.log('IS PREVIEW MODE');
-        matches = this.settings.paths.filter(
-          (path) =>
-            (path.scope === ClassPathScope.Preview || path.scope === ClassPathScope.Both) &&
-            activeView.file.path.startsWith(path.path)
-        );
-        container = this.getPreviewContainer(activeView);
-      } else if (this.isEditMode(activeView)) {
-        console.log('IS EDIT MODE');
-        matches = this.settings.paths.filter(
-          (path) =>
-            (path.scope === ClassPathScope.Edit || path.scope === ClassPathScope.Both) &&
-            activeView.file.path.startsWith(path.path)
-        );
-        container = this.getEditContainer(activeView);
+      // Get any linked views
+      let activeViews: MarkdownView[] = [activeView];
+      const leafGroup = this.app.workspace.getGroupLeaves((activeView.leaf as any).group);
+      if (leafGroup) {
+        activeViews = leafGroup
+          .map((leaf) => leaf.view)
+          .filter((view) => view instanceof MarkdownView) as MarkdownView[];
       }
-      console.log(matches);
-      console.log(container);
-      const classes: string[] = matches.flatMap((match) => match.classes);
-      console.log(classes);
-      this.applyClasses(classes, activeView, container);
+
+      // Remove and apply classes for each applicable view
+      activeViews.forEach((view) => {
+        this.removePreviousClasses(view);
+        let matches: ClassPath[] = [];
+        let container: Element;
+        if (this.isPreivewMode(view)) {
+          matches = this.settings.paths.filter(
+            (path) =>
+              (path.scope === ClassPathScope.Preview || path.scope === ClassPathScope.Both) &&
+              view.file.path.startsWith(path.path)
+          );
+          container = this.getPreviewContainer(view);
+        } else if (this.isEditMode(view)) {
+          matches = this.settings.paths.filter(
+            (path) =>
+              (path.scope === ClassPathScope.Edit || path.scope === ClassPathScope.Both) &&
+              view.file.path.startsWith(path.path)
+          );
+          container = this.getEditContainer(view);
+        }
+        const classes: string[] = matches.flatMap((match) => match.classes);
+        this.applyClasses(classes, view, container);
+      });
     }
   }
 
+  /**
+   * Save settings in plugin data
+   */
   saveSettings(): Promise<void> {
     return this.saveData(this.settings);
   }
 
+  /**
+   * Given a string of comma separated classnames,
+   * return them as an array
+   */
   getClassList(classString: string): string[] {
     return classString.split(',').map((cls) => cls.trim());
   }
 
+  /**
+   * Returns true if a view is in preview mode
+   */
   private isPreivewMode(view: MarkdownView): boolean {
     return (view.currentMode as any).type === 'preview';
   }
 
+  /**
+   * Returns true if a view is in edit/source mode
+   */
   private isEditMode(view: MarkdownView): boolean {
     return (view.currentMode as any).type === 'source';
   }
 
+  /**
+   * Add classes to an html element and store the
+   * applied classes along with a reference the view
+   * they were added to for removal later.
+   */
   private applyClasses(classes: string[], view: MarkdownView, container: Element): void {
     container.addClasses(classes);
     this.appliedClasses.push({ view, classes });
   }
 
+  /**
+   * Given a view, remove any extra classes this plugin
+   * has applied to that view.
+   */
   private removePreviousClasses(view: MarkdownView): void {
     const previewContainer = this.getPreviewContainer(view);
     const editContainer = this.getEditContainer(view);
@@ -93,6 +127,9 @@ export class AutoClassPlugin extends Plugin {
     this.appliedClasses = newApplied;
   }
 
+  /**
+   * Remove all applied classes from all views
+   */
   private removeAllClasses() {
     this.appliedClasses.forEach((applied) => {
       if (applied.view) {
@@ -108,10 +145,16 @@ export class AutoClassPlugin extends Plugin {
     });
   }
 
+  /**
+   * Get the element that preview classes are applied to
+   */
   private getPreviewContainer(view: MarkdownView): Element {
     return view.contentEl.querySelector('.markdown-preview-view');
   }
 
+  /**
+   * Get the element that edit/source classes are applied to
+   */
   private getEditContainer(view: MarkdownView): Element {
     return view.contentEl.querySelector('.markdown-source-view');
   }